refactor(extension): extract shared bridge transport interfaces (#682)
PanelMessageSender / PanelMessageSubscriber / OffscreenMessageHub were
declared identically in both follower-sprinkle-bridge.ts and
leader-sync-bridge.ts (same names, same shapes, two files). The
discriminateMsg helper was also duplicated. Any drift between the two
copies would silently let different test stubs satisfy different
bridges.

Extracted all four into bridge-transport.ts. Both bridge files import
from there and re-export the types so external callers don't see the
move.
